requirements.txt
flask==3.0.3
flask-cors==5.0.0
google-generativeai==0.7.2
PyPDF2==3.0.1
python-dotenv==1.0.1
gunicorn==21.2.0